+++ contracts/currency-network/CurrencyNetworkSafeMath.sol
@@ -1,42 +1,24 @@
 pragma solidity ^0.6.5;
 
-
 contract CurrencyNetworkSafeMath {
-
-    function safeSub(uint64 a, uint64 b)
-        internal
-        pure
-        returns (uint64)
-    {
+    function safeSub(uint64 a, uint64 b) internal pure returns (uint64) {
         require(b <= a, "SafeMath: Sub Overflow");
         return a - b;
     }
 
-    function safeAdd(uint64 a, uint64 b)
-        internal
-        pure
-        returns (uint64)
-    {
+    function safeAdd(uint64 a, uint64 b) internal pure returns (uint64) {
         uint64 c = a + b;
         require(c >= a, "SafeMath: Add Overflow");
         return c;
     }
 
-    function safeSubInt(int72 a, uint64 b)
-        internal
-        pure
-        returns (int72)
-    {
+    function safeSubInt(int72 a, uint64 b) internal pure returns (int72) {
         int72 c = a - b;
         require(c <= a, "SafeMath: SubInt Overflow");
         return c;
     }
 
-    function safeMinus(int72 a)
-        internal
-        pure
-        returns (int72)
-    {
+    function safeMinus(int72 a) internal pure returns (int72) {
         if (a == 0) {
             return a;
         }
@@ -45,20 +27,20 @@ contract CurrencyNetworkSafeMath {
         return c;
     }
 
-    function safeMinusInt256(int a)
-        internal
-        pure
-        returns (int)
-    {
+    function safeMinusInt256(int256 a) internal pure returns (int256) {
         if (a == 0) {
             return a;
         }
-        int c = -a;
+        int256 c = -a;
         require(c != a, "SafeMath: Minus Overflow");
         return c;
     }
 
-    function safeSumInt256(int a, int b) internal pure returns (int sum) {
+    function safeSumInt256(int256 a, int256 b)
+        internal
+        pure
+        returns (int256 sum)
+    {
         sum = a + b;
         if (a > 0 && b > 0) {
             require(sum > 0, "Overflow error.");
@@ -68,7 +50,7 @@ contract CurrencyNetworkSafeMath {
         }
     }
 
-    function checkIsNotMinInt256(int a) internal pure {
-        require(a != - 2 ** 255, "Prevent using value for minus overflow.");
+    function checkIsNotMinInt256(int256 a) internal pure {
+        require(a != -2**255, "Prevent using value for minus overflow.");
     }
 }
