+++ contracts/currency-network/DebtTracking.sol
@@ -2,22 +2,23 @@ pragma solidity ^0.6.5;
 
 import "./CurrencyNetworkSafeMath.sol";
 
-
 contract DebtTracking is CurrencyNetworkSafeMath {
-
     // mapping of a pair of user to the signed debt in the point of view of the lowest address
-    mapping (bytes32 => int) public debt;
+    mapping(bytes32 => int256) public debt;
 
-    event DebtUpdate(address _debtor, address _creditor, int _newDebt);
+    event DebtUpdate(address _debtor, address _creditor, int256 _newDebt);
 
     /**
      * @notice Used to increase the debt tracked by the currency network of msg.sender towards creditor address
      * @param creditor The address towards which msg.sender increases its debt
      * @param value The value to increase the debt by
      */
-    function increaseDebt(address creditor, uint value) external virtual {
-        int intValue = int(value);
-        require(uint(intValue) == value, "Value overflow, cannot be cast to int");
+    function increaseDebt(address creditor, uint256 value) external virtual {
+        int256 intValue = int256(value);
+        require(
+            uint256(intValue) == value,
+            "Value overflow, cannot be cast to int"
+        );
         _addToDebt(msg.sender, creditor, intValue);
     }
 
@@ -27,7 +28,12 @@ contract DebtTracking is CurrencyNetworkSafeMath {
      * @param creditor The address towards which the debtor owes money
      * @return the debt of the debtor to the creditor, equal to the opposite of the debt of the creditor to the debtor
      */
-    function getDebt(address debtor, address creditor) public view virtual returns (int) {
+    function getDebt(address debtor, address creditor)
+        public
+        view
+        virtual
+        returns (int256)
+    {
         if (debtor < creditor) {
             return debt[uniqueIdentifier(debtor, creditor)];
         } else {
@@ -35,28 +41,44 @@ contract DebtTracking is CurrencyNetworkSafeMath {
         }
     }
 
-    function _reduceDebt(address debtor, address creditor, uint value) internal {
-        int intValue = int(value);
-        require(uint(intValue) == value, "Value overflow, cannot be cast to int");
+    function _reduceDebt(
+        address debtor,
+        address creditor,
+        uint256 value
+    ) internal {
+        int256 intValue = int256(value);
+        require(
+            uint256(intValue) == value,
+            "Value overflow, cannot be cast to int"
+        );
         _addToDebt(debtor, creditor, safeMinusInt256(intValue));
     }
 
-    function _addToDebt(address debtor, address creditor, int value) internal {
-        int oldDebt = debt[uniqueIdentifier(debtor, creditor)];
+    function _addToDebt(
+        address debtor,
+        address creditor,
+        int256 value
+    ) internal {
+        int256 oldDebt = debt[uniqueIdentifier(debtor, creditor)];
         if (debtor < creditor) {
-            int newDebt = safeSumInt256(oldDebt, value);
+            int256 newDebt = safeSumInt256(oldDebt, value);
             checkIsNotMinInt256(newDebt);
             debt[uniqueIdentifier(debtor, creditor)] = newDebt;
             emit DebtUpdate(debtor, creditor, newDebt);
         } else {
-            int newDebt = safeSumInt256(oldDebt, safeMinusInt256(value));
+            int256 newDebt = safeSumInt256(oldDebt, safeMinusInt256(value));
             checkIsNotMinInt256(newDebt);
             debt[uniqueIdentifier(debtor, creditor)] = newDebt;
-            emit DebtUpdate(debtor, creditor, - newDebt);
+            emit DebtUpdate(debtor, creditor, -newDebt);
         }
     }
 
-    function uniqueIdentifier(address _a, address _b) internal pure virtual returns (bytes32) {
+    function uniqueIdentifier(address _a, address _b)
+        internal
+        pure
+        virtual
+        returns (bytes32)
+    {
         require(_a != _b, "Unique identifiers require different addresses");
         if (_a < _b) {
             return keccak256(abi.encodePacked(_a, _b));
