+++ contracts/tokens/DummyToken.sol
@@ -3,20 +3,17 @@ pragma solidity ^0.6.5;
 import "./../lib/SafeMath.sol";
 import "./../tokens/StandardToken.sol";
 
-
 contract DummyToken is StandardToken, SafeMath {
     string public name;
     string public symbol;
-    uint public decimals;
+    uint256 public decimals;
 
     constructor(
         string memory _name,
         string memory _symbol,
-        uint _decimals,
-        uint _totalSupply
-    )
-        public
-    {
+        uint256 _decimals,
+        uint256 _totalSupply
+    ) public {
         name = _name;
         symbol = _symbol;
         decimals = _decimals;
@@ -24,10 +21,8 @@ contract DummyToken is StandardToken, SafeMath {
         balances[msg.sender] = _totalSupply;
     }
 
-    function setBalance(address _target, uint _value)
-        public
-    {
-        uint currBalance = balanceOf(_target);
+    function setBalance(address _target, uint256 _value) public {
+        uint256 currBalance = balanceOf(_target);
         if (_value < currBalance) {
             totalSupply = safeSub(totalSupply, safeSub(currBalance, _value));
         } else {
