+++ contracts/tokens/Receiver_Interface.sol
@@ -1,12 +1,13 @@
 pragma solidity ^0.6.5;
 
-
 /*
-* Contract that is working with ERC223 tokens
-*/
+ * Contract that is working with ERC223 tokens
+ */
 
 interface ContractReceiver {
-
-    function tokenFallback(address _from, uint _value, bytes calldata _data) external;
-
+    function tokenFallback(
+        address _from,
+        uint256 _value,
+        bytes calldata _data
+    ) external;
 }
