+++ contracts/tokens/StandardToken.sol
@@ -2,12 +2,17 @@ pragma solidity ^0.6.5;
 
 import "./Token.sol";
 
-
 contract StandardToken is Token {
-
-    function transfer(address _to, uint _value) public override returns (bool) {
+    function transfer(address _to, uint256 _value)
+        public
+        override
+        returns (bool)
+    {
         //Default assumes totalSupply can't be over max (2^256 - 1).
-        if (balances[msg.sender] >= _value && balances[_to] + _value >= balances[_to]) {
+        if (
+            balances[msg.sender] >= _value &&
+            balances[_to] + _value >= balances[_to]
+        ) {
             balances[msg.sender] -= _value;
             balances[_to] += _value;
             emit Transfer(msg.sender, _to, _value);
@@ -17,8 +22,16 @@ contract StandardToken is Token {
         }
     }
 
-    function transferFrom(address _from, address _to, uint _value) public override returns (bool) {
-        if (balances[_from] >= _value && allowed[_from][msg.sender] >= _value && balances[_to] + _value >= balances[_to]) {
+    function transferFrom(
+        address _from,
+        address _to,
+        uint256 _value
+    ) public override returns (bool) {
+        if (
+            balances[_from] >= _value &&
+            allowed[_from][msg.sender] >= _value &&
+            balances[_to] + _value >= balances[_to]
+        ) {
             balances[_to] += _value;
             balances[_from] -= _value;
             allowed[_from][msg.sender] -= _value;
@@ -29,21 +42,30 @@ contract StandardToken is Token {
         }
     }
 
-    function balanceOf(address _owner) public view override returns (uint) {
+    function balanceOf(address _owner) public view override returns (uint256) {
         return balances[_owner];
     }
 
-    function approve(address _spender, uint _value) public override returns (bool) {
+    function approve(address _spender, uint256 _value)
+        public
+        override
+        returns (bool)
+    {
         allowed[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
         return true;
     }
 
-    function allowance(address _owner, address _spender) public view override returns (uint) {
+    function allowance(address _owner, address _spender)
+        public
+        view
+        override
+        returns (uint256)
+    {
         return allowed[_owner][_spender];
     }
 
-    mapping (address => uint) balances;
-    mapping (address => mapping (address => uint)) allowed;
-    uint public totalSupply;
+    mapping(address => uint256) balances;
+    mapping(address => mapping(address => uint256)) allowed;
+    uint256 public totalSupply;
 }
