+++ contracts/tokens/UnwETH.sol
@@ -17,29 +17,25 @@ pragma solidity ^0.6.5;
 
 import "../lib/Authorizable.sol";
 
-
 contract UnwEth is Authorizable {
-    string public name     = "Unwrapping Kovan Ether";
-    string public symbol   = "UKETH";
-    uint8  public decimals = 18;
+    string public name = "Unwrapping Kovan Ether";
+    string public symbol = "UKETH";
+    uint8 public decimals = 18;
 
-    event  Approval(address indexed src, address indexed guy, uint wad);
-    event  Transfer(address indexed src, address indexed dst, uint wad);
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
-    mapping (address => uint)                       public  balanceOf;
-    mapping (address => mapping (address => uint))  public  allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     receive() external payable {
         deposit();
     }
 
     // Todo Remove
-    function addAuthorizedAddress(address target)
-        public
-        override
-    {
+    function addAuthorizedAddress(address target) public override {
         super.addGlobalAuthorizedAddress(target);
     }
 
@@ -48,34 +44,40 @@ contract UnwEth is Authorizable {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf[msg.sender] >= wad);
         balanceOf[msg.sender] -= wad;
         msg.sender.transfer(wad);
         emit Withdrawal(msg.sender, wad);
     }
 
-    function totalSupply() public view returns (uint) {
+    function totalSupply() public view returns (uint256) {
         return address(this).balance;
     }
 
-    function approve(address guy, uint wad) public returns (bool) {
+    function approve(address guy, uint256 wad) public returns (bool) {
         allowance[msg.sender][guy] = wad;
         emit Approval(msg.sender, guy, wad);
         return true;
     }
 
-    function transfer(address payable dst, uint wad) public returns (bool) {
+    function transfer(address payable dst, uint256 wad) public returns (bool) {
         return transferFrom(msg.sender, dst, wad);
     }
 
-    function transferFrom(address src, address payable dst, uint wad)
-        public
-        returns (bool)
-    {
+    function transferFrom(
+        address src,
+        address payable dst,
+        uint256 wad
+    ) public returns (bool) {
         require(balanceOf[src] >= wad);
 
-        if (src != msg.sender && !globalAuthorized[msg.sender] && !authorizedBy[src][msg.sender] && allowance[src][msg.sender] != uint(-1)) {
+        if (
+            src != msg.sender &&
+            !globalAuthorized[msg.sender] &&
+            !authorizedBy[src][msg.sender] &&
+            allowance[src][msg.sender] != uint256(-1)
+        ) {
             require(allowance[src][msg.sender] >= wad);
             allowance[src][msg.sender] -= wad;
         }
@@ -88,7 +90,6 @@ contract UnwEth is Authorizable {
         return true;
     }
 }
-
 
 /*
                     GNU GENERAL PUBLIC LICENSE
